New account onboarding
and a bunch of tidying along the way
