+++ app/package.json
@@ -5,20 +5,20 @@
   "license": "MPL-2.0",
   "dependencies": {
     "@github/relative-time-element": "^4.3.0",
-    "@testing-library/react": "^14.0.0",
-    "@testing-library/user-event": "^14.5.1",
     "axios": "^1.5.1",
     "bootstrap": "^5.3.2",
     "formik": "^2.4.5",
     "humanize-duration": "^3.30.0",
     "luxon": "^3.4.3",
+    "primereact": "^9.6.2",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.8.0",
+    "react-bootstrap": "^2.9.0-beta.1",
     "react-bootstrap-icons": "^1.10.3",
     "react-dom": "^18.2.0",
     "react-router": "^6.16.0",
     "react-router-bootstrap": "^0.26.2",
     "react-router-dom": "^6.16.0",
+    "react-syntax-highlighter": "^15.5.0",
     "swagger-ui-react": "=5.7.2",
     "use-interval": "^1.4.0",
     "web-vitals": "^3.4.0"
@@ -51,6 +51,8 @@
   },
   "devDependencies": {
     "@testing-library/jest-dom": "^6.1.3",
+    "@testing-library/react": "^14.0.0",
+    "@testing-library/user-event": "^14.5.1",
     "@types/humanize-duration": "^3.27.1",
     "@types/jest": "^29.5.5",
     "@types/luxon": "^3.3.2",
@@ -61,6 +63,7 @@
     "@types/react-router": "^5.1.20",
     "@types/react-router-bootstrap": "^0.26.2",
     "@types/react-router-dom": "^5.3.3",
+    "@types/react-syntax-highlighter": "^15.5.7",
     "@types/swagger-ui-react": "^4.18.1",
     "@typescript-eslint/eslint-plugin": "^6.7.3",
     "@typescript-eslint/parser": "^6.7.3",
@@ -68,6 +71,7 @@
     "eslint": "^8.50.0",
     "eslint-config-prettier": "^9.0.0",
     "eslint-plugin-react": "^7.33.2",
+    "postcss-nesting": "^12.0.1",
     "prettier": "^3.0.3",
     "source-map-explorer": "^2.5.3",
     "typescript": "^5.2.2",
