+++ app/src/accounts/index.tsx
@@ -1,5 +1,5 @@
 import { RouteObject, defer, redirect } from "react-router-dom";
-import ApiClient, { PartialAccount } from "../ApiClient";
+import ApiClient, { NewAccount, UpdateAccount } from "../ApiClient";
 import AccountSummary from "./AccountSummary";
 import AccountForm from "./AccountForm";
 import AccountList from "./AccountList";
@@ -28,7 +28,7 @@ export default function accounts(
           switch (request.method) {
             case "POST": {
               const account = await apiClient.createAccount(
-                data as unknown as PartialAccount,
+                data as unknown as NewAccount,
               );
               return redirect(`/accounts/${account.id}`);
             }
@@ -39,29 +39,14 @@ export default function accounts(
       },
 
       {
-        path: ":account_id",
+        path: ":accountId",
         id: "account",
         loader({ params }) {
+          const { accountId } = params as { accountId: string };
           return defer({
-            account: apiClient.account(params.account_id as string),
+            account: apiClient.account(accountId),
           });
         },
-
-        async action({ params, request }) {
-          const data = Object.fromEntries(await request.formData());
-          switch (request.method) {
-            case "PATCH":
-              return {
-                account: await apiClient.updateAccount(
-                  params.account_id as string,
-                  data as unknown as PartialAccount,
-                ),
-              };
-            default:
-              throw new Error(`unexpected method ${request.method}`);
-          }
-        },
-
         shouldRevalidate(args) {
           return (
             typeof args.actionResult === "object" &&
@@ -69,8 +54,39 @@ export default function accounts(
             "account" in args.actionResult
           );
         },
+
+        async action({ params, request }) {
+          switch (request.method) {
+            case "PATCH": {
+              const data = await request.json();
+              return {
+                account: await apiClient.updateAccount(
+                  params.accountId as string,
+                  data as UpdateAccount,
+                ),
+              };
+            }
+            default:
+              throw new Error(`unexpected method ${request.method}`);
+          }
+        },
+
         children: [
-          { element: <AccountSummary />, path: "", index: true },
+          {
+            element: <AccountSummary />,
+            path: "",
+            index: true,
+            loader({ params }) {
+              const { accountId } = params as { accountId: string };
+              return defer({
+                apiTokens: apiClient.accountApiTokens(accountId),
+                tasks: apiClient.accountTasks(accountId),
+                hpkeConfigs: apiClient.accountHpkeConfigs(accountId),
+                aggregators: apiClient.accountAggregators(accountId),
+                account: apiClient.account(accountId),
+              });
+            },
+          },
           ...children,
         ],
       },
