+++ app/vite.config.ts
@@ -1,6 +1,7 @@
 import { defineConfig } from "vite";
 import eslint from "vite-plugin-eslint";
 import react from "@vitejs/plugin-react-swc";
+import postcssNesting from "postcss-nesting";
 
 export default defineConfig(() => ({
   build: {
@@ -11,5 +12,10 @@ export default defineConfig(() => ({
   server: {
     port: 8081,
     strictPort: true,
+  },
+  css: {
+    postcss: {
+      plugins: [postcssNesting],
+    },
   },
 }));
