+++ migration/src/lib.rs
@@ -18,6 +18,7 @@ mod m20230725_220134_add_vdafs_and_query_types_to_aggregators;
 mod m20230731_181722_rename_aggregator_bearer_token;
 mod m20230808_204859_create_hpke_config;
 mod m20230817_192017_add_protocol_to_aggregators;
+mod m20230907_175501_add_aggregator_onboarding_boolean_to_account;
 
 pub struct Migrator;
 
@@ -43,6 +44,7 @@ impl MigratorTrait for Migrator {
             Box::new(m20230731_181722_rename_aggregator_bearer_token::Migration),
             Box::new(m20230808_204859_create_hpke_config::Migration),
             Box::new(m20230817_192017_add_protocol_to_aggregators::Migration),
+            Box::new(m20230907_175501_add_aggregator_onboarding_boolean_to_account::Migration),
         ]
     }
 }
